Add scripts/make-deciles.sh and deprecate the other downloader.
