+++ .gitignore
@@ -11,5 +11,4 @@ processed_results
 analysis/
 
 notebooks/dev
-raw-data-google-ngram
 prompts
